refactor: cambiar todos los campos Double a BigDecimal en Vehiculo y Recorrido
- Vehiculo: combustible e indiceConsumo de Double a BigDecimal (precision 10, scale 2)
- Recorrido: consumo y litrosAbastecidos de Double a BigDecimal (precision 10, scale 2)
- DTOs actualizados: VehiculoRequest, VehiculoResponse, RecorridoRequest, RecorridoResponse
- RecorridoServiceImpl: toda la aritmetica reescrita con BigDecimal
  - Consumo: indiceConsumo.multiply(kilometros).divide(100, 2, HALF_UP)
  - Combustible: operaciones con .subtract(), .add(), .compareTo(ZERO)
  - Constantes CIEN y CERO como BigDecimal

+++ fleet-management/src/main/java/com/fleet/management/dto/recorrido/RecorridoRequest.java
@@ -3,6 +3,7 @@ package com.fleet.management.dto.recorrido;
 import jakarta.validation.constraints.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -24,7 +25,7 @@ public class RecorridoRequest {
     private Integer kilometros;
 
     @DecimalMin(value = "0.0", message = "Los litros abastecidos no pueden ser negativos")
-    private Double litrosAbastecidos;
+    private BigDecimal litrosAbastecidos;
 
     @Size(max = 50, message = "El numero de chip no puede exceder 50 caracteres")
     private String numeroChip;
